* Sign the video with my callsign * Add a few more logging messages * Change the time lapse from 24 to 36 hours

+++ animatemuf.py
@@ -12,7 +12,7 @@ from subprocess import Popen, PIPE
 from urllib.request import urlretrieve
 from datetime import datetime, timedelta
 
-from PIL import Image
+from PIL import Image, ImageFont, ImageDraw
 
 from sunslack import read_config
 
@@ -36,11 +36,12 @@ def retreive_files(config):
       target_name = os.path.join(config.target_dir, filename)
       if os.path.exists(target_name):
         continue
-      logging.info('Download %s', filename)
+      logging.info('%s saved', target_name)
       urlretrieve(NOAA + url['url'], target_name)
 
 def cleanup(config):
-  expire_time = datetime.utcnow() - timedelta(days=1, hours=1)
+  logging.info('Cleaning up old MUF images')
+  expire_time = datetime.utcnow() - timedelta(days=1, hours=12)
   for name in os.listdir(config.target_dir):
     if not name.startswith('CTIPe-MUF'):
       continue
@@ -54,28 +55,25 @@ def cleanup(config):
 
 def animate(config):
   # image sizes (1290, 700) (640, 400) (800, 600)
+  font = ImageFont.truetype(config.font, int(config.font_size))
   animation = os.path.join(config.target_dir, 'muf.gif')
   image_list = []
-  end = datetime.utcnow()
-  start = end - timedelta(hours=25)
 
   file_list = []
   for name in sorted(os.listdir(config.target_dir)):
-    if not name.startswith('CTIPe-MUF'):
+    if not name.startswith('CTIPe-MUF_'):
       continue
-    try:
-      file_time = datetime.strptime(RE_TIME(name).group(1), '%Y%m%dT%H%M%S')
-    except AttributeError:
-      continue
-    if start < file_time < end:
-      file_list.append(name)
+    file_list.append(name)
 
+  logging.info('Processing: %d images', len(file_list))
   for name in file_list:
     fullname = os.path.join(config.target_dir, name)
     logging.debug('Add %s', name)
     image = Image.open(fullname)
     image = image.convert('RGB')
     image = image.resize((800, 600), Image.Resampling.LANCZOS)
+    draw = ImageDraw.Draw(image)
+    draw.text((25, 555), "W6BSD - MUF 36 hours animation", font=font, fill="gray")
     image_list.append(image)
 
   if len(image_list) > 2:
@@ -88,15 +86,19 @@ def animate(config):
 def gen_video(config):
   logfile = os.path.join(config.target_dir, 'muf.log')
   gif_file = os.path.join(config.target_dir, 'muf.gif')
-  cmd = f'{config.converter} {gif_file} {config.video_file}'
 
+  if not os.path.isfile(config.converter):
+    logging.error('Video converter %s not found', config.converter)
+    return
+
+  cmd = f'{config.converter} {gif_file} {config.video_file}'
   with open(logfile, "w") as err:
     print(cmd, file=err)
     proc = Popen(cmd.split(), shell=False, stdout=PIPE, stderr=err)
   logging.info(f"Saving %s video file", config.video_file)
   proc.wait()
   if proc.returncode != 0:
-    logging.error('Error generating the video file')
+    logging.error('Error generating the video file. Status code: %d', proc.returncode)
 
 def main(args=sys.argv[:1]):
   parser = argparse.ArgumentParser(description='MUF animation')
@@ -108,6 +110,10 @@ def main(args=sys.argv[:1]):
   _config = read_config(os.path.expanduser(opts.config))
   config = _config['animatemuf']
   del _config
+
+  if not os.path.isdir(config.target_dir):
+    logging.error("The target directory %s does not exist", config.target_dir)
+    return
 
   retreive_files(config)
   cleanup(config)
